requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
requests>=2.28.0
face_recognition>=1.3.0
numpy>=1.21.0
playwright>=1.40.0
pillow>=10.0.0
python-dotenv>=1.0.0
git+https://github.com/NickWaterton/samsung-tv-ws-api.git#egg=samsungtvws[async,encrypted]
